refactor logic to make more readable
